Move benchmarks directory to src
