+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "start": "./scripts/benchmarkSeparately.sh",
     "start:dirty": "INITIAL_STATE=dirty ./scripts/benchmarkSeparately.sh",
-    "benchmark": "node benchmarks/index.js",
+    "benchmark": "node src/benchmarks/index.js",
     "prebuild": "rm -rf dist",
     "build": "tsc -p tsconfig.build.json",
     "test": "mocha",
