+++ src/benchmarks/02-erc20-transfers.js
@@ -1,7 +1,7 @@
 const { ContractFactory, utils } = require('ethers')
-const { measureExecution } = require("./utils/measureExecution")
-const { choose1 } = require('./utils/random')
-const ERC20Mock = require('../contracts/ERC20Mock.json')
+const { measureExecution } = require('./utils/measureExecution')
+const { choose2, randomEthValue } = require('./utils/random')
+const ERC20Mock = require('../../contracts/ERC20Mock.json')
 
 exports.run = async function (runs, provider) {
   const wallets = provider.getWallets()
@@ -15,13 +15,19 @@ exports.run = async function (runs, provider) {
   const data = new Array(runs)
 
   for (let i = 0; i < runs; i++) {
-    data[i] = choose1(wallets)
+    const [from, to] = choose2(wallets)
+    const value = randomEthValue(1, 15)
+    data[i] = {
+      from,
+      to,
+      value
+    }
   }
 
-  return measureExecution(async function() {
+  return measureExecution(async function () {
     for (let i = 0; i < runs; i++) {
-      const holder = data[i]
-      await token.balanceOf(holder.address)
+      const { from, to, value } = data[i]
+      await token.connect(from).transfer(to.address, value)
     }
   })
 }
