+++ src/benchmarks/04-erc20-calls.js
@@ -1,7 +1,7 @@
 const { ContractFactory, utils } = require('ethers')
 const { measureExecution } = require("./utils/measureExecution")
-const { choose1, randomEthAddress, randomEthValue } = require('./utils/random')
-const ERC20Mock = require('../contracts/ERC20Mock.json')
+const { choose1 } = require('./utils/random')
+const ERC20Mock = require('../../contracts/ERC20Mock.json')
 
 exports.run = async function (runs, provider) {
   const wallets = provider.getWallets()
@@ -15,17 +15,13 @@ exports.run = async function (runs, provider) {
   const data = new Array(runs)
 
   for (let i = 0; i < runs; i++) {
-    data[i] = {
-      holder: choose1(wallets),
-      spender: randomEthAddress(),
-      value: randomEthValue(10, 50)
-    }
+    data[i] = choose1(wallets)
   }
 
-  return measureExecution(async function () {
+  return measureExecution(async function() {
     for (let i = 0; i < runs; i++) {
-      const { holder, spender, value } = data[i]
-      await token.connect(holder).approve(spender, value)
+      const holder = data[i]
+      await token.balanceOf(holder.address)
     }
   })
 }
